Merged PR 12526350: Dxe Readiness Capture/Validation Tool
Dxe Readiness Capture/Validation Tool

This PR serves as the foundation for the final dxe readiness capture/validation tool.
Currently, it only includes basic logging functionality. Once this PR is merged, we will eventually
add other components, such as a global allocator and JSON dumping functionality in a follow up PRs.
This repository acts as a placeholder for the tool's development until we find
the right place for it in ODP.

#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces the initial commit for the DXE Readiness Validation Tool, which builds and runs as a Rust binary in QEMU firmware.
- Added `build_and_run_platform_handoff_validation_tool.py` to build, patch, and run the Rust DXE Core binary.
- Updated `README.md` with detailed instructions on building, testing, and publishing the tool.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->