Clarify PTL Support
Consolidates details in Makefile.toml. Adds a `build-intel-ptl` task
to build Intel Panther Lake capture binaries. At this time, the actual
binary content is the same, so the binary crate is renamed to a common
name but separate build commands are used so the build can accommodate
generation-specific changes if needed in the future.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>